feat: add support for configFile encryption.
