+++ polaris-configuration/polaris-configuration-client/src/main/java/com/tencent/polaris/configuration/client/internal/ConfigFileManager.java
@@ -21,7 +21,7 @@ import com.tencent.polaris.api.config.configuration.CryptoConfig;
 import com.tencent.polaris.api.plugin.common.PluginTypes;
 import com.tencent.polaris.api.plugin.configuration.ConfigFileConnector;
 import com.tencent.polaris.api.plugin.configuration.ConfigFileResponse;
-import com.tencent.polaris.api.plugin.crypto.ConfigFilterCrypto;
+import com.tencent.polaris.api.plugin.filter.ConfigFileFilter;
 import com.tencent.polaris.client.api.SDKContext;
 import com.tencent.polaris.configuration.api.core.ConfigFile;
 import com.tencent.polaris.configuration.api.core.ConfigFileFormat;
@@ -42,7 +42,7 @@ public class ConfigFileManager {
 
     private SDKContext context;
 
-    private ConfigFilterCrypto crypto;
+    private ConfigFileFilter crypto;
 
     private ConfigFileConnector connector;
 
@@ -70,7 +70,7 @@ public class ConfigFileManager {
         String cryptoType = sdkContext.getConfig().getConfigFile().getConfigFilterConfig().getChain().get(0);
         CryptoConfig pluginConfig = sdkContext.getConfig().getConfigFile().getConfigFilterConfig().getPluginConfig(cryptoType, CryptoConfig.class);
         this.context = sdkContext;
-        this.crypto = (ConfigFilterCrypto) sdkContext.getExtensions().getPlugins()
+        this.crypto = (ConfigFileFilter) sdkContext.getExtensions().getPlugins()
                 .getPlugin(PluginTypes.CONFIG_FILTER_CRYPTO.getBaseType(), pluginConfig.getType());
         this.connector = (ConfigFileConnector) sdkContext.getExtensions().getPlugins()
                 .getPlugin(PluginTypes.CONFIG_FILE_CONNECTOR.getBaseType(), configFileConnectorType);
