+++ polaris-plugins/polaris-plugin-api/src/main/java/com/tencent/polaris/api/plugin/common/PluginTypes.java
@@ -22,7 +22,7 @@ import com.tencent.polaris.api.plugin.cache.FlowCache;
 import com.tencent.polaris.api.plugin.circuitbreaker.CircuitBreaker;
 import com.tencent.polaris.api.plugin.circuitbreaker.InstanceCircuitBreaker;
 import com.tencent.polaris.api.plugin.configuration.ConfigFileConnector;
-import com.tencent.polaris.api.plugin.crypto.ConfigFilterCrypto;
+import com.tencent.polaris.api.plugin.filter.ConfigFileFilter;
 import com.tencent.polaris.api.plugin.detect.HealthChecker;
 import com.tencent.polaris.api.plugin.loadbalance.LoadBalancer;
 import com.tencent.polaris.api.plugin.location.LocationProvider;
@@ -100,7 +100,7 @@ public enum PluginTypes {
     /**
      * 配置加密扩展点
      */
-    CONFIG_FILTER_CRYPTO(new PluginType(ConfigFilterCrypto.class, 2)),
+    CONFIG_FILTER_CRYPTO(new PluginType(ConfigFileFilter.class, 2)),
 
     /**
      * 配置文件加载器扩展点
