+++ polaris-plugins/polaris-plugins-configfilefilter/configfilefilter-crypto/src/main/resources/META-INF/services/com.tencent.polaris.api.plugin.filter.ConfigFileFilter
@@ -1,0 +1,1 @@
+com.tencent.polaris.plugins.configfilefilter.CryptoConfigFileFilter
