+++ polaris-plugins/polaris-plugins-configfilefilter/pom.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>polaris-plugins</artifactId>
+        <groupId>com.tencent.polaris</groupId>
+        <version>${revision}</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>polaris-plugins-configfilefilter</artifactId>
+
+    <name>Polaris Plugins Configfilter</name>
+    <description>Polaris Plugins Configfilter POM</description>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>configfilefilter-crypto</module>
+    </modules>
+
+
+</project>
