--- polaris-plugins/polaris-plugins-configfilter/src/main/resources/META-INF/services/com.tencent.polaris.api.plugin.crypto.ConfigFilterCrypto
@@ -1,1 +1,0 @@
-com.tencent.polaris.plugins.configfilter.ConfigFileFilter
