+++ polaris-plugins/pom.xml
@@ -28,6 +28,6 @@
         <module>polaris-plugins-cache</module>
         <module>polaris-plugins-configuration-connector</module>
 		<module>polaris-plugins-location</module>
-        <module>polaris-plugins-configfilter</module>
+        <module>polaris-plugins-configfilefilter</module>
 	</modules>
 </project>
